Fix #8393.
Former-commit-id: b693fed75402ca51bc7fd0399b6038e91dcda023

+++ source/ch/cyberduck/core/openstack/SwiftHpUrlProvider.java
@@ -63,7 +63,7 @@ public class SwiftHpUrlProvider extends SwiftUrlProvider {
     }
 
     @Override
-    protected DescriptiveUrlBag sign(final Region region, final Path file, final int expiry) {
+    protected DescriptiveUrlBag sign(final Region region, final Path file, final long expiry) {
         final String path = region.getStorageUrl(
                 containerService.getContainer(file).getName(), containerService.getKey(file)).getRawPath();
         final Credentials credentials = session.getHost().getCredentials();
